refactor(rust): split OCR utilities
