+++ litellm-rust/crates/core/src/ocr/document_intelligence.rs
@@ -1,0 +1,110 @@
+use std::time::{Duration, Instant};
+
+use serde_json::Value;
+
+use crate::CoreResult;
+use crate::error::CoreError;
+use crate::http_utils::truncate_error_body;
+
+use super::client::http_client;
+
+const POLL_TIMEOUT_SECS: u64 = 120;
+
+fn same_origin(left: &str, right: &str) -> bool {
+    let Ok(left) = reqwest::Url::parse(left) else {
+        return false;
+    };
+    let Ok(right) = reqwest::Url::parse(right) else {
+        return false;
+    };
+    left.scheme() == right.scheme()
+        && left.host_str() == right.host_str()
+        && left.port_or_known_default() == right.port_or_known_default()
+}
+
+fn retry_after_secs(response: &reqwest::Response) -> u64 {
+    response
+        .headers()
+        .get(reqwest::header::RETRY_AFTER)
+        .and_then(|value| value.to_str().ok())
+        .and_then(|value| value.parse::<u64>().ok())
+        .unwrap_or(2)
+}
+
+fn operation_status(response_json: &Value) -> CoreResult<&str> {
+    let status = response_json
+        .get("status")
+        .and_then(Value::as_str)
+        .ok_or(CoreError::MissingField("status"))?;
+    match status {
+        "succeeded" => Ok("succeeded"),
+        "running" | "notStarted" => Ok("running"),
+        "failed" => {
+            let message = response_json
+                .get("error")
+                .and_then(|error| error.get("message"))
+                .and_then(Value::as_str)
+                .unwrap_or("Unknown error");
+            Err(CoreError::InvalidResponse(format!(
+                "Azure Document Intelligence analysis failed: {message}"
+            )))
+        }
+        other => Err(CoreError::InvalidResponse(format!(
+            "Unknown operation status: {other}"
+        ))),
+    }
+}
+
+pub(super) async fn poll_document_intelligence(
+    operation_url: &str,
+    original_url: &str,
+    headers: &[(String, String)],
+    timeout: Option<Duration>,
+) -> CoreResult<Value> {
+    if !same_origin(operation_url, original_url) {
+        return Err(CoreError::InvalidResponse(
+            "Azure Document Intelligence: rejected cross-origin polling URL".to_string(),
+        ));
+    }
+
+    let start = Instant::now();
+    let timeout = timeout.unwrap_or(Duration::from_secs(POLL_TIMEOUT_SECS));
+    loop {
+        if start.elapsed() > timeout {
+            return Err(CoreError::Network(format!(
+                "Azure Document Intelligence operation polling timed out after {} seconds",
+                timeout.as_secs()
+            )));
+        }
+
+        let mut request_builder = http_client().get(operation_url);
+        for (key, value) in headers {
+            if key.eq_ignore_ascii_case("ocp-apim-subscription-key") {
+                request_builder = request_builder.header(key, value);
+            }
+        }
+        let response = request_builder
+            .send()
+            .await
+            .map_err(|err| CoreError::Network(err.to_string()))?;
+        let retry_after = retry_after_secs(&response);
+        let status = response.status();
+        let text = response
+            .text()
+            .await
+            .map_err(|err| CoreError::Network(err.to_string()))?;
+        if !status.is_success() {
+            return Err(CoreError::Http {
+                status: status.as_u16(),
+                body: truncate_error_body(&text),
+            });
+        }
+        let response_json: Value = serde_json::from_str(&text).map_err(|err| {
+            CoreError::InvalidResponse(format!("invalid Azure DI poll response JSON: {err}"))
+        })?;
+        if operation_status(&response_json)? == "succeeded" {
+            return Ok(response_json);
+        }
+        tokio::time::sleep(Duration::from_secs(retry_after)).await;
+    }
+}
