+++ litellm-rust/crates/core/src/ocr/handler.rs
@@ -1,10 +1,11 @@
 use crate::CoreResult;
 use crate::error::CoreError;
+use crate::http_utils::truncate_error_body;
 use crate::ocr::transformation::OcrResponseHandling;
 use serde_json::Value;
 
 use super::client::http_client;
-use super::common_utils::{poll_document_intelligence, truncate_error_body};
+use super::document_intelligence::poll_document_intelligence;
 use super::types::ProviderOcrRequest;
 
 pub(crate) async fn execute_ocr_provider_call(request: ProviderOcrRequest) -> CoreResult<Value> {
