+++ litellm-rust/crates/core/src/ocr/provider_config.rs
@@ -1,0 +1,30 @@
+use crate::ocr::transformation::OcrProviderConfig;
+use crate::providers::azure_ai::ocr::transformation::{
+    AZURE_AI_OCR_CONFIG, AZURE_DOCUMENT_INTELLIGENCE_OCR_CONFIG,
+};
+use crate::providers::mistral::ocr::transformation::MISTRAL_OCR_CONFIG;
+use crate::providers::vertex_ai::ocr::transformation as vertex_ai;
+use crate::providers::vertex_ai::ocr::transformation::{
+    VERTEX_AI_DEEPSEEK_OCR_CONFIG, VERTEX_AI_OCR_CONFIG,
+};
+
+pub(super) fn ocr_provider_config(
+    provider: &str,
+    model: &str,
+) -> Option<&'static dyn OcrProviderConfig> {
+    match provider {
+        "mistral" => Some(&MISTRAL_OCR_CONFIG),
+        "azure_ai" if is_azure_document_intelligence_model(model) => {
+            Some(&AZURE_DOCUMENT_INTELLIGENCE_OCR_CONFIG)
+        }
+        "azure_ai" => Some(&AZURE_AI_OCR_CONFIG),
+        "vertex_ai" if vertex_ai::is_deepseek_model(model) => Some(&VERTEX_AI_DEEPSEEK_OCR_CONFIG),
+        "vertex_ai" => Some(&VERTEX_AI_OCR_CONFIG),
+        _ => None,
+    }
+}
+
+fn is_azure_document_intelligence_model(model: &str) -> bool {
+    let model = model.to_ascii_lowercase();
+    model.contains("doc-intelligence") || model.contains("documentintelligence")
+}
